# Optimize re-rendering child workflows.