This optimization moves the management of child workflows out
of `Behavior` and `LifetimeTracker` and directly into `SubtreeManager`.
It uses roughly the same algorithm as the Swift implementation (I believe),
and during a render pass just effectively tags nodes as still in-use or not,
by moving them between two linked lists. After the render pass, everything that
is in the old list is torn down and the lists are swapped for the next render pass.

This reduces the number of allocations required by each `WorkflowNode`'s `render`,
and in the optimal case of a workflow just re-rendering all its children, has
almost no overhead.


[Spreadsheet](https://docs.google.com/spreadsheets/d/1e4OBW0RmKscQyQPIYMLHf0Tzj2WbBnOndVLnEcJ5MAg/edit?usp=sharing)

Benchmark | (treeShape) | Mode | Cnt | Score (before) | Error | Score (after) | Error | Units | Speedup
-- | -- | -- | -- | -- | -- | -- | -- | -- | --
WorkflowNodeBenchmark.renderPassAlternateLeaves | DEEP | avgt | 3 | 310.463 | 7.559 | 72.753 | 4.480 | us/op | 4.27
WorkflowNodeBenchmark.renderPassAlternateLeaves | BUSHY | avgt | 3 | 563.374 | 50.840 | 445.627 | 119.474 | us/op | 1.26
WorkflowNodeBenchmark.renderPassAlternateLeaves | SQUARE | avgt | 3 | 504.244 | 8.309 | 343.108 | 124.959 | us/op | 1.47
WorkflowNodeBenchmark.renderPassAlternateOneLeaf | DEEP | avgt | 3 | 315.172 | 18.100 | 71.529 | 8.432 | us/op | 4.41
WorkflowNodeBenchmark.renderPassAlternateOneLeaf | BUSHY | avgt | 3 | 294.026 | 22.039 | 67.544 | 0.996 | us/op | 4.35
WorkflowNodeBenchmark.renderPassAlternateOneLeaf | SQUARE | avgt | 3 | 273.824 | 5.861 | 54.579 | 1.845 | us/op | 5.02
WorkflowNodeBenchmark.renderPassAlternateWorkers | DEEP | avgt | 3 | 323.121 | 20.304 | 69.888 | 6.516 | us/op | 4.62
WorkflowNodeBenchmark.renderPassAlternateWorkers | BUSHY | avgt | 3 | 556.811 | 80.929 | 444.579 | 100.674 | us/op | 1.25
WorkflowNodeBenchmark.renderPassAlternateWorkers | SQUARE | avgt | 3 | 502.989 | 7.628 | 344.899 | 46.005 | us/op | 1.46
WorkflowNodeBenchmark.renderPassAlwaysLeaves | DEEP | avgt | 3 | 312.049 | 93.199 | 65.652 | 3.975 | us/op | 4.75
WorkflowNodeBenchmark.renderPassAlwaysLeaves | BUSHY | avgt | 3 | 292.758 | 19.453 | 61.276 | 0.459 | us/op | 4.78
WorkflowNodeBenchmark.renderPassAlwaysLeaves | SQUARE | avgt | 3 | 269.063 | 11.410 | 53.762 | 7.107 | us/op | 5.00